**IMPORTANT** MUST READ!
FTX seems to have changed the way futures are listed.

With Equilibrium, **TRX-PERP** becomes **TRX/USD:USD**

It would appear that FTX is moving towards allowing perpetuals to be in
more than one quote currrency. This appears to be an exchange level
change. This could also be CCXT unifying FTX nomenclature to be in line
with most other exchanges.

There are still some cases where -PERP still works, but i'm not sure for
how long.

*Remember, when you run the update script, it makes sure **everything** is properly updated.*

Changes to be committed:
	modified:   Base/Library/JRRapi.py
	modified:   Base/Library/JRRconfig.py
	modified:   Base/Library/JRRsupport.py
	modified:   Extras/Equilibrium/Equilibrium.future
	modified:   Extras/Equilibrium/Equilibrium.spot