Bug fixes.
ListMarkets now displays spread and 1 day volatility equation improved.

Since FTX went out of the way to claim there was no advantage to limit
orders over market orders because spread ate up the difference in
profits, I added the ability to detect and block high spreads.

The following demonstrates how to limit spread. Orders are NOT allowed if
spread is ABOVE the specified value. This value is expressed in points of
the difference between bid and ask.

In your order,

"Spread":"0.25",

In your config file (applies to ALL orders on a given exchange),

"SpreadOverride":"0.25"

Changes to be committed:
	modified:   Base/CCXT-PlaceOrder.future
	modified:   Base/CCXT-PlaceOrder.spot
	modified:   Base/Library/JRRconfig.py
	modified:   Base/Library/JRRledger.py
	modified:   Extras/ListMarkets
	modified:   install
	modified:   update