Base work to support OANDA in place.
Config files now define framework. Curretly supported: CCXT, OANDA.

Stubs in place for any number of future frameworks. This approach does NOT
try to unify the frameworks as the differences are too significant. The
Relay framework IS trying to provide an overlay class to blur the
differences as reasonable possible.

Changes to be committed:
	modified:   Base/Library/JRRapi.py
	new file:   Base/Library/JRRoanda.py
	modified:   Base/Library/JackrabbitRelay.py
	modified:   requirements.txt